@chaosagent 2023-08-25
The use-case is to have an easy way of seeing how changes affect the whole suite of tests, for example when changing optimizer.py. I find this useful, but you don't have to merge this if you don't think it is useful.

I think the current CI setup is good-enough.