@wozeparrot 2023-08-25
So I kinda do want this, it does seem kinda useful, but for me the csv output is wayyyy more useful for `tinymod` rather than parsing the text output.

Does this change existing behaviour? like with the way it's run in CI right now will this still work the same way?