@chaosagent 2023-08-25
If you don't mkdir reports, it will behave exactly the same as before.