@chaosagent 2023-08-25
I changed it so that comparison against baseline only happens on RPT=2.