@chaosagent 2023-08-31
> I don't like that test_speed_v_torch now imports from extra, makes it harder to understand.

Would you prefer this be put into test/? Or do you feel that this is too much complexity?

The idea was to make these reports for all sorts of different tests. We should have single kernel speed tests for all major kernels in all the models we support, and also generate these benchmarks for the model tests.