@wozeparrot 2023-09-04
Is it possible to have `perf_report.py` be a wrapper for `test_speed_v_torch.py` or the other benchmarks? Tho that probably just adds more complexity. Maybe for this we need something like what's needed for `tinyboard`? some sort of generic logging hook system.